logging: Rename files and test from log_msg2 to log_msg
Renaming log_msg2 files to log_msg.
