+++ include/zephyr/logging/log_backend.h
@@ -6,7 +6,7 @@
 #ifndef ZEPHYR_INCLUDE_LOGGING_LOG_BACKEND_H_
 #define ZEPHYR_INCLUDE_LOGGING_LOG_BACKEND_H_
 
-#include <zephyr/logging/log_msg2.h>
+#include <zephyr/logging/log_msg.h>
 #include <stdarg.h>
 #include <zephyr/sys/__assert.h>
 #include <zephyr/sys/util.h>
