+++ include/zephyr/logging/log_backend_std.h
@@ -6,7 +6,7 @@
 #ifndef ZEPHYR_LOG_BACKEND_STD_H_
 #define ZEPHYR_LOG_BACKEND_STD_H_
 
-#include <zephyr/logging/log_msg2.h>
+#include <zephyr/logging/log_msg.h>
 #include <zephyr/logging/log_output.h>
 #include <zephyr/kernel.h>
