+++ include/zephyr/logging/log_ctrl.h
@@ -8,7 +8,7 @@
 
 #include <zephyr/kernel.h>
 #include <zephyr/logging/log_backend.h>
-#include <zephyr/logging/log_msg2.h>
+#include <zephyr/logging/log_msg.h>
 #include <zephyr/logging/log_internal.h>
 
 #ifdef __cplusplus
