+++ subsys/logging/log_msg.c
