+++ tests/subsys/logging/log_msg/CMakeLists.txt
