+++ tests/subsys/logging/log_msg/prj.conf
